Moving mapping functions to MappingF
