+++ src/Jeebs.Data/Functions/MappingF.cs
@@ -1,0 +1,14 @@
+﻿// Jeebs Rapid Application Development
+// Copyright (c) bfren.uk - licensed under https://mit.bfren.uk/2013
+
+namespace F.DataF
+{
+	/// <summary>
+	/// Mapping Functions
+	/// </summary>
+	public static partial class MappingF
+	{
+		/// <summary>Messages</summary>
+		public static partial class Msg { }
+	}
+}
