+++ tests/Tests.Jeebs.Data/Functions/MappingF/GetColumnWithAttribute_Tests.cs
@@ -1,11 +1,15 @@
 ﻿// Jeebs Unit Tests
 // Copyright (c) bfren.uk - licensed under https://mit.bfren.uk/2013
 
+using Jeebs;
 using Jeebs.Data.Entities;
+using Jeebs.Data.Mapping;
+using Jeebs.Data.Mapping.Mapper_Tests;
 using Xunit;
-using static Jeebs.Data.Mapping.Mapper.Msg;
+using static F.DataF.MappingF;
+using static F.DataF.MappingF.Msg;
 
-namespace Jeebs.Data.Mapping.Mapper_Tests
+namespace F.DataF.MappingF_Tests
 {
 	public class GetColumnWithAttribute_Tests
 	{
@@ -13,10 +17,10 @@ namespace Jeebs.Data.Mapping.Mapper_Tests
 		public void Missing_Id_Property_Attribute_Returns_None_With_NoPropertyWithAttributeMsg()
 		{
 			// Arrange
-			var columns = Mapper.GetMappedColumns<FooWithoutIdAttribute>(new FooTable()).UnsafeUnwrap();
+			var columns = GetMappedColumns<FooWithoutIdAttribute>(new FooTable()).UnsafeUnwrap();
 
 			// Act
-			var result = Mapper.GetColumnWithAttribute<FooWithoutIdAttribute, IdAttribute>(columns);
+			var result = GetColumnWithAttribute<FooWithoutIdAttribute, IdAttribute>(columns);
 
 			// Assert
 			var none = result.AssertNone();
@@ -27,10 +31,10 @@ namespace Jeebs.Data.Mapping.Mapper_Tests
 		public void Multiple_Id_Properties_Returns_None_With_TooManyPropertiesWithAttributeMsg()
 		{
 			// Arrange
-			var columns = Mapper.GetMappedColumns<FooWithMultipleIdAttributes>(new FooTable()).UnsafeUnwrap();
+			var columns = GetMappedColumns<FooWithMultipleIdAttributes>(new FooTable()).UnsafeUnwrap();
 
 			// Act
-			var result = Mapper.GetColumnWithAttribute<FooWithMultipleIdAttributes, IdAttribute>(columns);
+			var result = GetColumnWithAttribute<FooWithMultipleIdAttributes, IdAttribute>(columns);
 
 			// Assert
 			var none = result.AssertNone();
