+++ tests/Tests.Jeebs.Data/Functions/MappingF/ValidateTable_Tests.cs
@@ -2,6 +2,7 @@
 // Copyright (c) bfren.uk - licensed under https://mit.bfren.uk/2013
 
 using Xunit;
+using static F.DataF.MappingF;
 
 namespace Jeebs.Data.Mapping.Mapper_Tests
 {
@@ -14,7 +15,7 @@ namespace Jeebs.Data.Mapping.Mapper_Tests
 			var e0 = $"The definition of table '{typeof(FooTableWithoutBar0).FullName}' is missing field '{nameof(Foo.Bar0)}'.";
 
 			// Act
-			var (valid, errors) = Mapper.ValidateTable<Foo>(new FooTableWithoutBar0());
+			var (valid, errors) = ValidateTable<Foo>(new FooTableWithoutBar0());
 
 			// Assert
 			Assert.False(valid);
@@ -30,7 +31,7 @@ namespace Jeebs.Data.Mapping.Mapper_Tests
 			var e2 = $"The definition of table '{typeof(FooTableWithoutAny).FullName}' is missing field '{nameof(Foo.Bar1)}'.";
 
 			// Act
-			var (valid, errors) = Mapper.ValidateTable<Foo>(new FooTableWithoutAny());
+			var (valid, errors) = ValidateTable<Foo>(new FooTableWithoutAny());
 
 			// Assert
 			Assert.False(valid);
@@ -48,7 +49,7 @@ namespace Jeebs.Data.Mapping.Mapper_Tests
 			var e0 = $"The definition of entity '{typeof(Foo).FullName}' is missing property '{nameof(FooTableWithBar2.Bar2)}'.";
 
 			// Act
-			var (valid, errors) = Mapper.ValidateTable<Foo>(new FooTableWithBar2());
+			var (valid, errors) = ValidateTable<Foo>(new FooTableWithBar2());
 
 			// Assert
 			Assert.False(valid);
@@ -64,7 +65,7 @@ namespace Jeebs.Data.Mapping.Mapper_Tests
 			var e2 = $"The definition of entity '{typeof(Foo).FullName}' is missing property '{nameof(FooTableWithBar234.Bar4)}'.";
 
 			// Act
-			var (valid, errors) = Mapper.ValidateTable<Foo>(new FooTableWithBar234());
+			var (valid, errors) = ValidateTable<Foo>(new FooTableWithBar234());
 
 			// Assert
 			Assert.False(valid);
