+++ tests/Tests.Jeebs.Data/Functions/QueryF/GetModelProperties_Tests.cs
